requirements.txt
# Core dependencies
requests>=2.31,<3
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# API dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# OpenAI (required for OpenAI embeddings and LLM)
openai>=1.0.0

# Local embeddings (required for local provider)
sentence-transformers>=2.2.0
torch>=2.0.0
